rcan

Trait Capability

pub trait Capability: Serialize {
    // Required method
    fn permits(&self, other: &Self) -> bool;
}

A trait for types that define a capability.

Capabilities can be compared using Capability::permits, which determines whether one capability grants permission to perform another.

A common implementation of this trait might be an enum representing different RPC request types.

The Capability type must be serializable so it can be included in the signature payload in an Rcan.

Required Methods

fn permits(&self, other: &Self) -> bool

Determines if self permits other.

Returns true if self grants permission to perform the other capability, otherwise returns false.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.